+++ matisse/__init__.py
@@ -1,6 +1,7 @@
 """Package defining the MaTiSSe.py program."""
 
-__version__ = "1.4.1"
+__appname__ = "MaTiSSe.py"
+__version__ = "1.4.2"
 __author__ = "Stefano Zaghi"
 __author_email__ = "stefano.zaghi@gmail.com"
 __license__ = "GNU General Public License v3 (GPLv3)"